requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.1.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tablib==3.5.0
uritemplate==4.1.1
